@@ -1,9 +1,13 @@
+//go:build license
+
 package maiao
 
 import (
-	"io/ioutil"
+	"math"
+	"os"
 	"strings"
 	"testing"
+	"time"
 
 	pkggodevclient "github.com/guseggert/pkggodev-client"
 	"github.com/stretchr/testify/require"
@@ -21,28 +25,62 @@ var (
 		"MPL-2.0":      {},
 	}
 
-	knownUndectedLicenses = map[string]string{
+	knownUndetectedLicenses = map[string]string{
 		// bufpipe was later added the MIT license: https://github.com/acomagu/bufpipe/blob/cd7a5f79d3c413d14c0c60fd31dae7b397fc955a/LICENSE
 		"github.com/acomagu/bufpipe@v1.0.3": "MIT",
 	}
+
+	maxRetryAttempts = 4
 )
 
+func calculateBackoffTime(attempt int) time.Duration {
+	return time.Duration(math.Pow(5, float64(attempt)))*time.Second + 10*time.Second
+}
+
+func isRateLimitError(err error) bool {
+	return strings.Contains(err.Error(), "Too Many Requests")
+}
+
 func TestLicenses(t *testing.T) {
-	b, err := ioutil.ReadFile("go.mod")
+	b, err := os.ReadFile("go.mod")
 	require.NoError(t, err)
 	file, err := modfile.Parse("go.mod", b, nil)
 	require.NoError(t, err)
 	client := pkggodevclient.New()
 	for _, req := range file.Require {
-		pkg, err := client.DescribePackage(pkggodevclient.DescribePackageRequest{
-			Package: req.Mod.Path,
-		})
+		var pkg *pkggodevclient.Package
+		var err error
+
+		attempt := 0
+		for ; attempt < maxRetryAttempts; attempt++ {
+			pkg, err = client.DescribePackage(pkggodevclient.DescribePackageRequest{
+				Package: req.Mod.Path,
+			})
+			if err == nil {
+				break
+			}
+
+			if !isRateLimitError(err) {
+				break
+			}
+
+			if attempt < maxRetryAttempts-1 {
+				waitTime := calculateBackoffTime(attempt)
+				t.Logf("Rate limited while checking %s, waiting %v before retry %d/%d", req.Mod.Path, waitTime, attempt+1, maxRetryAttempts)
+				time.Sleep(waitTime)
+			}
+		}
+
+		if attempt >= maxRetryAttempts {
+			t.Logf("Skipping license check for %s due to persistent rate limiting", req.Mod.Path)
+		}
+
 		require.NoError(t, err)
 		licences := strings.Split(pkg.License, ",")
 		for _, license := range licences {
 			license = strings.TrimSpace(license)
 			if license == "None detected" {
-				if known, ok := knownUndectedLicenses[req.Mod.String()]; ok {
+				if known, ok := knownUndetectedLicenses[req.Mod.String()]; ok {
 					license = known
 				}
 			}
